Create GE areospace Task 1: Build a user interface.html
For this task, you'll be building the user interface in Vue SFC Playground.

 Navigate to this site and familiarise yourself with the interface. You'll be making edits to the autogenerated App.vue file.
Edit the App.vue file to add a heading that reads "Compute Product".
 Embed an image of your choice.
Add variables for the two numbers to be multiplied together. Then, add text to the page that displays these numbers.
Add a compute button. Configure it so that when the button is pressed, the button disappears and is replaced by the text indicating the product of the two numbers.